Add nested regions for states
